+++ src/main/kotlin/blackjack/GameController.kt
@@ -1,14 +1,14 @@
 package blackjack
 
 import blackjack.domain.BlackJackGame
-import blackjack.domain.Card
 import blackjack.domain.Deck
 import blackjack.domain.Player
-import blackjack.domain.PlayerResults
-import blackjack.view.BlackJackInputView
-import blackjack.view.BlackJackOutputView
+import blackjack.domain.card.Card
+import blackjack.domain.gameResult.PlayerResults
 import blackjack.view.InputView
 import blackjack.view.OutputView
+import blackjack.view.blackjackView.BlackJackInputView
+import blackjack.view.blackjackView.BlackJackOutputView
 
 class GameController(
     private val inputView: InputView,
