+++ src/main/kotlin/blackjack/domain/card/Rank.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 enum class Rank(val score: Int) {
     ACE(1),
