+++ src/main/kotlin/blackjack/domain/gameResult/GameResultStatus.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.gameResult
 
 enum class GameResultStatus {
     PLAYER_WIN,
