+++ src/main/kotlin/blackjack/domain/gameResult/PlayerResult.kt
@@ -1,3 +1,5 @@
-package blackjack.domain
+package blackjack.domain.gameResult
+
+import blackjack.domain.Player
 
 data class PlayerResult(val player: Player, val status: GameResultStatus)
