+++ src/main/kotlin/blackjack/view/blackjackView/BlackJackInputView.kt
@@ -1,4 +1,4 @@
-package blackjack.view
+package blackjack.view.blackjackView
 
 import java.lang.IllegalArgumentException
