+++ src/test/kotlin/blackjack/PlayerResultsTest.kt
@@ -1,12 +1,12 @@
 package blackjack
 
-import blackjack.domain.Card
 import blackjack.domain.Dealer
-import blackjack.domain.GameResultStatus
 import blackjack.domain.Player
-import blackjack.domain.PlayerResults
-import blackjack.domain.Rank
-import blackjack.domain.Suit
+import blackjack.domain.card.Card
+import blackjack.domain.card.Rank
+import blackjack.domain.card.Suit
+import blackjack.domain.gameResult.GameResultStatus
+import blackjack.domain.gameResult.PlayerResults
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
